Switch solution type to String
Day 5 requires a string solution. I wasn't sure if that would happen
when making the initial framework so I went with a large number type,
but unfortunately that has to change. Fourtunately it's happened early;
should be easy to change the return type of the `part{1,2}` methods and
slap a `.to_string()` onto your solutions.